feat(ideaview) : idea view endpoints changed
