+++ api/ideaview/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from .views import *
+
+urlpatterns = [
+    path('total/', TotalIdeaCountAPI.as_view()),
+    path('csv/', ImportOrgCSVAPI.as_view()),
+    path('list/', IdeaCountListAPI.as_view()),
+]
