+++ api/organization/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import OrganizationAPI, AssignOrganizationAPI, OrgVisitedAPI, OrganizationListAPI, ImportOrgCSVAPI, OrganizationIdeaCountAPI, UserIdeaViewCountAPI
+from .views import *
 
 urlpatterns = [
     path('', OrganizationAPI.as_view()),
@@ -7,7 +7,7 @@ urlpatterns = [
     path('assign/', AssignOrganizationAPI.as_view()),
     path('update-status/', OrgVisitedAPI.as_view()),
     path('list/', OrganizationListAPI.as_view()),
-    path('idea/csv/',ImportOrgCSVAPI.as_view()),
-    path('idea/total/',OrganizationIdeaCountAPI.as_view()),
-    path('idea/users/',UserIdeaViewCountAPI.as_view()),
+    # path('idea/csv/',ImportOrgCSVAPI.as_view()),
+    # path('idea/total/',OrganizationIdeaCountAPI.as_view()),
+    # path('idea/users/',UserIdeaViewCountAPI.as_view()),
 ]
